tests/wpt/web-platform-tests/css/selectors/focus-visible-006-manual.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
  <title>CSS Test (Selectors): contenteditable elements always match :focus-visible</title>
  <link rel="author" title="Alice Boxhall" href="aboxhall@chromium.org" />
  <link rel="help" href="https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <style>
    span[contenteditable] {
        border: 1px solid black;
        background-color: white;
        padding: 2px 5px;
    }

    :root {
      --focus-background: hsl(50, 94%, 72%);
      --focus-ring-color: rgb(59, 153, 252);
    }

    :focus-visible {
      outline: var(--focus-ring-color) auto 5px;
    }

    * {
      outline: 0;
    }

    :focus {
      background-color: var(--focus-background);
    }
  </style>
</head>
<body>
  This test checks that <code>:focus-visible</code> always matches on elements with <code>contenteditable=true</code> set.
  <ol id="instructions">
    <li>If the user-agent does not claim to support the <code>:focus-visible</code> pseudo-class then SKIP this test.</li>
    <li><strong>Click</strong> the content editable span below to focus it.</li>
    <li>If the element does not have a blue outline, then the test result is FAILURE. If the element has a blue outline, then the test result is SUCCESS.</li>
  </ol>
  <br />
  <div>
    <span id="el" contenteditable>Focus me</span>
  </div>
  <script>
    async_test(function(t) {
      el.addEventListener("focus", t.step_func(function() {
        assert_equals(getComputedStyle(el).outlineColor, "rgb(59, 153, 252)");
        t.done();
      }));
    }, "Focus should always match :focus-visible on content editable divs");
  </script>
</body>
</html>
